SessionService
A SessionService has been developed and it will be in charge of storing the authentication information of every logged in user in the application, using a GenServer for it. Tokens has been deleted from the application since they are no longer needed, and also some functions have been deleted or renamed in Repo and Sign_in mostly to use the new service.

This closes #5